del prof

+++ service/api/delete-profile.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"errors"
 	"net/http"
+	"strconv"
 
 	"git.sapienzaapps.it/fantasticcoffee/fantastic-coffee-decaffeinated/service/api/reqcontext"
 	"git.sapienzaapps.it/fantasticcoffee/fantastic-coffee-decaffeinated/service/database"
@@ -12,25 +13,34 @@ import (
 
 func (rt *_router) deleteProfile(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
 	var deletedUser Users
+	id, err := strconv.Atoi("id")
+	authToken := r.Header.Get("authToken")
 
-	err := json.NewDecoder(r.Body).Decode(&deletedUser)
-	if err != nil {
+	bool, err := rt.db.CheckAuthToken(id, authToken)
+	if bool == true {
+		err := json.NewDecoder(r.Body).Decode(&deletedUser)
+		if err != nil {
 
-		return
-	}
+			return
+		}
 
-	err = rt.db.DeleteProfile(deletedUser.ToDatabaseUser())
-	if errors.Is(err, database.ErrUserDoesNotExist) {
+		err = rt.db.DeleteProfile(deletedUser.ToDatabaseUser())
+		if errors.Is(err, database.ErrUserDoesNotExist) {
 
-		w.WriteHeader(http.StatusNotFound)
-		return
-	} else if err != nil {
+			w.WriteHeader(http.StatusNotFound)
+			return
+		} else if err != nil {
 
-		ctx.Logger.WithError(err).WithField("UserID", deletedUser).Error("can't delete profile")
+			ctx.Logger.WithError(err).WithField("UserID", deletedUser).Error("can't delete profile")
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+
+		}
+
+		w.WriteHeader(http.StatusNoContent)
+	} else {
+		ctx.Logger.WithError(err).Error("Uncorrect token")
 		w.WriteHeader(http.StatusInternalServerError)
 		return
-
 	}
-
-	w.WriteHeader(http.StatusNoContent)
 }
